requirements.txt
amqp==5.3.1
asgiref==3.10.0
bidict==0.23.1
billiard==4.2.2
blinker==1.9.0
brotli==1.2.0
celery==5.5.3
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.7.1
cryptography==45.0.7
defusedxml==0.7.1
distlib==0.4.0
Django==5.2.8
django-cors-headers==4.9.0
django-debug-toolbar==6.0.0
django-filter==25.1
django-templated-mail==1.1.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
docopt==0.6.2
drf-nested-routers==0.95.0
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
flower==2.0.1
gevent==25.9.1
geventhttpclient==2.3.5
greenlet==3.2.4
h11==0.16.0
humanize==4.14.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
locust==2.42.3
locust-cloud==1.28.1
MarkupSafe==3.0.3
model-bakery==1.20.5
msgpack==1.1.2
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
pipenv==2025.0.4
platformdirs==4.5.0
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psutil==7.1.3
psycopg==3.2.10
psycopg-binary==3.2.10
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
pytest-watch==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.3
python-socketio==5.14.3
python3-openid==3.2.0
pytz==2025.2
pyzmq==27.1.0
redis==7.0.1
requests==2.32.4
requests-oauthlib==2.0.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tornado==6.5.2
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
virtualenv==20.34.0
watchdog==6.0.0
wcwidth==0.2.14
websocket-client==1.9.0
Werkzeug==3.1.3
wsproto==1.3.1
zope.event==6.1
zope.interface==8.1.1
